requirements.txt
Flask==2.3.2
gunicorn==21.2.0
google-api-python-client==2.126.0
pymongo==4.6.1
requests==2.31.0
google-auth==2.29.0
google-cloud-aiplatform==1.50.0
google-genai
python-dotenv==1.0.1
google-generativeai
pymongo 
numpy 
scipy